chore(evm-block-recovery): routines::scan_evm_state_roots review fixes 2

+++ evm-utils/evm-block-recovery/src/cli.rs
@@ -246,7 +246,7 @@ pub struct ScanEvmStateRootsArgs {
     pub evm_state_path: PathBuf,
 
     #[arg(short, long)]
-    pub workers: u8,
+    pub workers: u16,
 
     #[arg(short, long)]
     pub secondary: bool,
